feat: rename transfer2 to transfer
